Allow scope references in interpolated trees
This is perfectly fine now that the interpolation process follows the
expression tree, as the referenced values are embedded in the tree.
In theory this should work on the evaluation side as well with the
caveat that we would have to be able to cast the value referenced in the
tree to the required type (and implement the expression binding).