Refactor InterceptorGenerator as an analyzer
Refactored the expression interpolation interceptor generator as a
simple analyzer. This was a tough decision given the amount of work
poured into the generator; but the ~2x performance gain (in the vast
majority of cases) did not justify the enormous complexity, associated
brittleness, and number of edge cases in which the interceptor
implementation was not possible.